Apollo: Release source code for 55.4 (1504)

+++ libwallet/domain/model/security_card/security_card_metadata.go
@@ -4,7 +4,7 @@ type SecurityCardMetadata struct {
 	GlobalPublicKeyInHex string
 	CardVendorInHex      string
 	CardModelInHex       string
-	FirmwareVersion      int
-	UsageCount           int
+	FirmwareVersion      uint16
+	UsageCount           uint16
 	LanguageCodeInHex    string
 }
