+++ libwallet/service/model/register_security_card_json.go
@@ -3,7 +3,7 @@ package model
 type RegisterSecurityCardJson struct {
 	CardPublicKeyInHex   string                   `json:"cardPublicKeyInHex"`
 	ClientPublicKeyInHex string                   `json:"clientPublicKeyInHex"`
-	PairingSlot          int                      `json:"pairingSlot"`
+	PairingSlot          uint16                   `json:"pairingSlot"`
 	Metadata             SecurityCardMetadataJson `json:"metadata"`
 	MacInHex             string                   `json:"macInHex"`
 	GlobalSignCardInHex  string                   `json:"globalSignCardInHex"`
